fix: address Bedrock guardrail review feedback
